Move helpers to appropriate folder
